requirements.txt
absl-py==2.0.0
aiobotocore==2.7.0
aiohttp==3.8.6
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
albumentations==1.1.0
amqp==5.1.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
asyncssh==2.14.0
atpublic==4.0
attrs==23.1.0
Babel==2.13.1
billiard==4.1.0
black==23.10.1
bleach==6.1.0
blessed==1.20.0
boto3==1.28.64
botocore==1.31.64
brotlipy==0.7.0
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
clearml==1.13.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
conda==23.3.1
conda-build==3.24.0
configobj==5.0.8
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.8.0
dnspython==2.3.0
docker-pycreds==0.4.0
dpath==2.1.6
dulwich==0.21.6
dvc==3.28.0
dvc-data==2.20.0
dvc-http==2.30.2
dvc-objects==1.1.0
dvc-render==0.6.0
dvc-s3==2.23.0
dvc-studio-client==0.15.0
dvc-task==0.3.0
entrypoints==0.4
exceptiongroup==1.1.1
expecttest==0.1.4
fastjsonschema==2.18.1
flake8==6.1.0
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
funcy==2.0
furl==2.1.3
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.1.0
gpustat==1.1.1
grandalf==0.8
grpcio==1.59.0
gto==1.5.0
huggingface-hub==0.18.0
hydra-core==1.3.2
hypothesis==6.75.2
imageio==2.31.6
imgaug==0.4.0
iniconfig==2.0.0
iopath==0.1.9
ipykernel==6.26.0
ipython-genutils==0.2.0
isoduration==20.11.0
iterative-telemetry==0.0.8
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.8.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kiwisolver==1.4.5
kombu==5.3.2
lazy_loader==0.3
loguru==0.7.2
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
matplotlib==3.8.0
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.6
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
mypy==1.6.1
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
nvidia-ml-py==12.535.108
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
orderedmultidict==1.0.1
orjson==3.9.10
overrides==7.4.0
pandas==2.1.2
pandocfilters==1.5.0
pathlib2==2.3.7.post1
pathspec==0.11.2
pathtools==0.1.2
Pillow==9.4.0
platformdirs==3.11.0
portalocker==2.8.2
prometheus-client==0.18.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pydot==1.4.2
pyflakes==3.1.0
pygit2==1.13.2
pygtrie==2.5.0
PyJWT==2.4.0
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python-etcd==0.4.5
python-json-logger==2.0.7
pyzmq==25.1.1
qudida==0.0.4
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
s3fs==2023.10.0
s3transfer==0.7.0
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
scmrepo==1.4.1
seaborn==0.13.0
semver==3.0.2
Send2Trash==1.8.2
sentry-sdk==1.33.1
setproctitle==1.3.3
shortuuid==1.0.11
shtab==1.6.4
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sqltrie==0.8.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.8
tinycss2==1.2.1
tomlkit==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchelastic==0.2.2
torchtext==0.15.2
torchvision==0.15.2
tornado==6.3.3
triton==2.0.0
typer==0.9.0
types-dataclasses==0.6.6
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
vine==5.0.0
voluptuous==0.13.1
wandb==0.15.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.0
wrapt==1.15.0
yacs==0.1.8
yarl==1.9.2
zc.lockfile==3.0.post1
